Bug Fixes
-Start on moon again
-Included powerup timers in robot
-Magnet strength depends on timer
-Being invincible no longer stops time :sweat_smile:
-Energy now dies after being collected (instead of off screen)
-Powerups now die, instead of living forever
-Timers now add instead of being overwritten
-Magnet adjusted
-Debris particles larger

+++ include/energy.h
@@ -17,7 +17,7 @@ class energy : public game_object{
     void logic(int newMotion, robot *ourRobot);
 
     // Move towards robot
-    void move_towards( int mov_x, int mov_y, int speed);
+    void move_towards( float mov_x, float mov_y, float speed);
 };
 
 #endif
